refactor 502cr to be in its own calculator
