+++ spec/lib/efile/md/md502_calculator_spec.rb
@@ -10,241 +10,6 @@ describe Efile::Md::Md502Calculator do
     )
   end
 
-  describe "#calculate_md502_cr_part_b_line_3" do
-    before do
-      intake.direct_file_data.fed_agi = agi
-      instance.calculate
-    end
-
-    context "when filer is non-mfj" do
-      let(:filing_status) { "single" }
-
-      context "the agi is negative" do
-        let(:agi) { -10 }
-        it 'returns the correct decimal value' do
-          expect(instance.lines[:MD502CR_PART_B_LINE_3].value).to eq(0.32)
-        end
-      end
-
-      context "the agi is 0" do
-        let(:agi) { 0 }
-        it 'returns the correct decimal value' do
-          expect(instance.lines[:MD502CR_PART_B_LINE_3].value).to eq(0.32)
-        end
-      end
-
-      context "the agi is 90,001" do
-        let(:agi) { 90_001 }
-        it 'returns the correct decimal value' do
-          expect(instance.lines[:MD502CR_PART_B_LINE_3].value).to eq(0.2208)
-        end
-      end
-
-      context "the agi is $103,651" do
-        let(:agi) { 103_651 }
-        it 'returns the correct decimal value' do
-          expect(instance.lines[:MD502CR_PART_B_LINE_3].value).to eq(0.1984e0)
-        end
-      end
-
-      context "the agi is over $107,001" do
-        let(:agi) { 107_001 }
-        it 'returns the correct decimal value' do
-          expect(instance.lines[:MD502CR_PART_B_LINE_3].value).to eq(0.1856e0)
-        end
-      end
-    end
-
-    context "when filer is mfj" do
-      let(:intake) { create(:state_file_md_intake, :with_senior_spouse) }
-
-      context 'the agi is $62,001' do
-        let(:agi) { 62_001 }
-        it 'returns the correct decimal value' do
-          expect(instance.lines[:MD502CR_PART_B_LINE_3].value).to eq(0.3040)
-        end
-      end
-
-      context 'the agi is $167,1001' do
-        let(:agi) { 167_101 }
-        it 'returns the correct decimal value' do
-          expect(instance.lines[:MD502CR_PART_B_LINE_3].value).to eq(0.192e0)
-        end
-      end
-    end
-  end
-
-  describe '#calculate_md502_cr_part_b_line_4' do
-    let(:filing_status) { "single" }
-
-    before do
-      intake.direct_file_data.fed_agi = 10
-      intake.direct_file_data.fed_credit_for_child_and_dependent_care_amount = 10
-      instance.calculate
-    end
-
-    it 'returns the correct decimal value' do
-      expect(instance.lines[:MD502CR_PART_B_LINE_4].value).to eq(3)
-    end
-  end
-
-  describe '#calculate_md502_cr_part_m_line_1' do
-    context "when filing status is mfj, qss or hoh" do
-      let(:filing_status) { "married_filing_jointly" }
-      context "when filer is 65 or older" do
-        before do
-          intake.primary_birth_date = Date.new(MultiTenantService.statefile.current_tax_year - 65, 1, 1)
-        end
-        context "when spouse is 65 or older" do
-          before do
-            intake.spouse_birth_date = Date.new(MultiTenantService.statefile.current_tax_year - 65, 1, 1)
-          end
-          context "when agi <= $150,000" do
-            before do
-              intake.direct_file_data.fed_agi = 150_000
-              instance.calculate
-            end
-            it "awards a credit of $1750" do
-              expect(instance.lines[:MD502CR_PART_M_LINE_1].value).to eq(1_750)
-            end
-          end
-          context "when agi > $150,000" do
-            before do
-              intake.direct_file_data.fed_agi = 150_001
-              instance.calculate
-            end
-            it "awards no credit" do
-              expect(instance.lines[:MD502CR_PART_M_LINE_1].value).to eq(0)
-            end
-          end
-        end
-        context "when spouse is under 65" do
-          before do
-            intake.spouse_birth_date = Date.new(MultiTenantService.statefile.current_tax_year - 64, 1, 1)
-          end
-          context "when agi <= $150,000" do
-            before do
-              intake.direct_file_data.fed_agi = 150_000
-              instance.calculate
-            end
-            it "awards a credit of $1000" do
-              expect(instance.lines[:MD502CR_PART_M_LINE_1].value).to eq(1_000)
-            end
-          end
-          context "when agi > $150,000" do
-            before do
-              intake.direct_file_data.fed_agi = 150_001
-              instance.calculate
-            end
-            it "awards no credit" do
-              expect(instance.lines[:MD502CR_PART_M_LINE_1].value).to eq(0)
-            end
-          end
-        end
-      end
-      context "when filer is under 65" do
-        before do
-          intake.primary_birth_date = Date.new(MultiTenantService.statefile.current_tax_year - 64, 1, 1)
-        end
-        context "when spouse is 65 or older" do
-          before do
-            intake.spouse_birth_date = Date.new(MultiTenantService.statefile.current_tax_year - 65, 1, 1)
-          end
-          context "when agi <= $150,000" do
-            before do
-              intake.direct_file_data.fed_agi = 150_000
-              instance.calculate
-            end
-            it "awards a credit of $1000" do
-              expect(instance.lines[:MD502CR_PART_M_LINE_1].value).to eq(1_000)
-            end
-          end
-          context "when agi > $150,000" do
-            before do
-              intake.direct_file_data.fed_agi = 150_001
-              instance.calculate
-            end
-            it "awards no credit" do
-              expect(instance.lines[:MD502CR_PART_M_LINE_1].value).to eq(0)
-            end
-          end
-        end
-        context "when spouse is under 65" do
-          before do
-            intake.spouse_birth_date = Date.new(MultiTenantService.statefile.current_tax_year - 64, 1, 1)
-          end
-          context "when agi <= $150,000" do
-            before do
-              intake.direct_file_data.fed_agi = 150_000
-              instance.calculate
-            end
-            it "awards no credit" do
-              expect(instance.lines[:MD502CR_PART_M_LINE_1].value).to eq(0)
-            end
-          end
-          context "when agi > $150,000" do
-            before do
-              intake.direct_file_data.fed_agi = 150_001
-              instance.calculate
-            end
-            it "awards no credit" do
-              expect(instance.lines[:MD502CR_PART_M_LINE_1].value).to eq(0)
-            end
-          end
-        end
-      end
-    end
-    context "when filing status is single or mfs" do
-      let(:filing_status) { "single" }
-      context "when filer is 65 or older" do
-        before do
-          intake.primary_birth_date = Date.new(MultiTenantService.statefile.current_tax_year - 65, 1, 1)
-        end
-        context "when agi <= $100,000" do
-          before do
-            intake.direct_file_data.fed_agi = 100_000
-            instance.calculate
-          end
-          it "awards a credit of $1750" do
-            expect(instance.lines[:MD502CR_PART_M_LINE_1].value).to eq(1_000)
-          end
-        end
-        context "when agi > $100,000" do
-          before do
-            intake.direct_file_data.fed_agi = 100_001
-            instance.calculate
-          end
-          it "awards no credit" do
-            expect(instance.lines[:MD502CR_PART_M_LINE_1].value).to eq(0)
-          end
-        end
-      end
-      context "when filer is under 65" do
-        before do
-          intake.primary_birth_date = Date.new(MultiTenantService.statefile.current_tax_year - 64, 1, 1)
-        end
-        context "when agi <= $100,000" do
-          before do
-            intake.direct_file_data.fed_agi = 100_000
-            instance.calculate
-          end
-          it "awards no credit" do
-            expect(instance.lines[:MD502CR_PART_M_LINE_1].value).to eq(0)
-          end
-        end
-        context "when agi > $100,000" do
-          before do
-            intake.direct_file_data.fed_agi = 100_001
-            instance.calculate
-          end
-          it "awards no credit" do
-            expect(instance.lines[:MD502CR_PART_M_LINE_1].value).to eq(0)
-          end
-        end
-      end
-    end
-  end
-
   describe "#calculate_line_a_primary" do
     context 'primary not claimed as a dependent' do
       before do
